+++ packages/lit-html/rollup-version-stability-test.config.js
@@ -1,0 +1,12 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+import {defaultConfig} from './rollup.config.js';
+
+export default defaultConfig({
+  outputDir: './version-stability-build/',
+  testPropertyPrefix: 'VERSION_TEST_',
+});