+++ packages/lit-html/src/test/lit-html_test.ts
@@ -241,6 +241,29 @@ suite('lit-html', () => {
       assertRender(html`<script></script>${'A'}`, '<script></script>A');
     });
 
+    test('text after script element with binding', () => {
+      assertRender(
+        html`<script type="foo">${'A'}</script>${'B'}`,
+        '<script type="foo">A</script>B'
+      );
+      assertRender(
+        html`<script type="foo">1${'A'}</script>${'B'}`,
+        '<script type="foo">1A</script>B'
+      );
+      assertRender(
+        html`<script type="foo">${'A'}1</script>${'B'}`,
+        '<script type="foo">A1</script>B'
+      );
+      assertRender(
+        html`<script type="foo">${'A'}${'B'}</script>${'C'}`,
+        '<script type="foo">AB</script>C'
+      );
+      assertRender(
+        html`<script type="foo">${'A'}</script><p>${'B'}</p>`,
+        '<script type="foo">A</script><p>B</p>'
+      );
+    });
+
     test('text after style element', () => {
       assertRender(html`<style></style>${'A'}`, '<style></style>A');
     });