+++ packages/localize-tools/testdata/build-transform-xliff/goldens/tsout/zh_CN/foo.js
@@ -3,7 +3,7 @@
  * Copyright 2020 Google LLC
  * SPDX-License-Identifier: BSD-3-Clause
  */
-import {LitElement, html} from 'lit-element';
+import {LitElement, html} from 'lit';
 const {getLocale} = {getLocale: () => 'zh_CN'};
 console.log(`Locale is ${getLocale()}`);
 window.addEventListener('lit-localize-status', (event) => {
@@ -36,7 +36,18 @@ html`[SALT] Hello <b>World</b>!`;
 html`Hello <b><!-- comment -->World</b>!`;
 // Custom ID
 ('Hello World');
-export class MyElement extends LitElement {
+// updateWhenLocaleChanges -> undefined
+export class MyElement1 extends LitElement {
+  constructor() {
+    super();
+    undefined;
+  }
+  render() {
+    return html`<p>你好 <b>世界</b>! (${getLocale()})</p>`;
+  }
+}
+// @localized -> removed
+export class MyElement2 extends LitElement {
   render() {
     return html`<p>你好 <b>世界</b>! (${getLocale()})</p>`;
   }