+++ packages/localize-tools/testdata/extract-xlb-fresh/goldens/foo.ts
@@ -4,14 +4,14 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {html} from 'lit-html';
+import {html} from 'lit';
 import {msg} from '@lit/localize';
 
 msg('Hello World!', {
-  desc: 'Description of Hello World'
+  desc: 'Description of Hello World',
 });
 
 const name = 'friend';
 msg(html`Hello <b>${name}!</b>`, {
-  desc: 'Description of Hello $(name)'
+  desc: 'Description of Hello $(name)',
 });